Code Samples using newer Pytorch Training Tooklit

+++ pyt_breast_cancer.py
@@ -6,70 +6,74 @@ My experiments with Python, Machine Learning & Deep Learning.
 This code is meant for education purposes only & is not intended for commercial/production use!
 Use at your own risk!! I am not responsible if your CPU or GPU gets fried :D
 """
-# %% https://code.visualstudio.com/docs/python/jupyter-support-py
 import warnings
 
 warnings.filterwarnings('ignore')
 
 import os
-import sys
-import random
 import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
 import seaborn as sns
+from sklearn.metrics import classification_report, confusion_matrix
 
 # tweaks for libraries
-np.set_printoptions(precision=6, linewidth=1024, suppress=True)
+np.set_printoptions(precision = 6, linewidth = 1024, suppress = True)
 plt.style.use('seaborn')
-sns.set(style='darkgrid', context='notebook', font_scale=1.10)
+sns.set(style = 'darkgrid', context = 'notebook', font_scale = 1.10)
 
 # Pytorch imports
 import torch
 
 print('Using Pytorch version: ', torch.__version__)
 import torch.nn as nn
-import torch.nn.functional as F
-
+from torchmetrics.classification import (
+    BinaryAccuracy, BinaryF1Score,
+    BinaryAUROC
+)
 # import pytorch_toolkit - training Nirvana :)
-import pytorch_toolkit as pytk
+import pytorch_training_toolkit as t3
 
 # to ensure that you get consistent results across runs & machines
 # @see: https://discuss.pytorch.org/t/reproducibility-over-different-machines/63047
-seed = 123
-random.seed(seed)
-os.environ['PYTHONHASHSEED'] = str(seed)
-np.random.seed(seed)
-torch.manual_seed(seed)
-if torch.cuda.is_available():
-    torch.cuda.manual_seed(seed)
-    torch.cuda.manual_seed_all(seed)
-    torch.backends.cudnn.deterministic = True
-    torch.backends.cudnn.benchmark = False
-    torch.backends.cudnn.enabled = False
-
-data_file_path = './data/wisconsin_breast_cancer.csv'
+SEED = t3.seed_all()
+data_file_path = os.path.join(
+    os.path.dirname(__file__), 'data', 'wisconsin_breast_cancer.csv'
+)
+DEVICE = torch.device("cuda:0") \
+    if torch.cuda.is_available() else torch.device("cpu")
 
 
-def download_data_file():
-    url = 'https://archive.ics.uci.edu/ml/machine-learning-databases/breast-cancer-wisconsin/wdbc.data'
+def download_data_file(force = False):
+    if (not os.path.exists(data_file_path)) or force:
+        url = 'https://archive.ics.uci.edu/ml/machine-learning-databases/breast-cancer-wisconsin/wdbc.data'
 
-    df_cols = [
-        "id", "diagnosis", "radius_mean", "texture_mean", "perimeter_mean", "area_mean",
-        "smoothness_mean", "compactness_mean", "concavity_mean", "concave points_mean",
-        "symmetry_mean", "fractal_dimension_mean", "radius_se", "texture_se", "perimeter_se",
-        "area_se", "smoothness_se", "compactness_se", "concavity_se", "concave points_se",
-        "symmetry_se", "fractal_dimension_se", "radius_worst", "texture_worst", "perimeter_worst",
-        "area_worst", "smoothness_worst", "compactness_worst", "concavity_worst",
-        "concave points_worst", "symmetry_worst", "fractal_dimension_worst",
-    ]
+        df_cols = [
+            "id", "diagnosis", "radius_mean", "texture_mean", "perimeter_mean",
+            "area_mean",
+            "smoothness_mean", "compactness_mean", "concavity_mean",
+            "concave points_mean",
+            "symmetry_mean", "fractal_dimension_mean", "radius_se",
+            "texture_se",
+            "perimeter_se",
+            "area_se", "smoothness_se", "compactness_se", "concavity_se",
+            "concave points_se",
+            "symmetry_se", "fractal_dimension_se", "radius_worst",
+            "texture_worst",
+            "perimeter_worst",
+            "area_worst", "smoothness_worst", "compactness_worst",
+            "concavity_worst",
+            "concave points_worst", "symmetry_worst", "fractal_dimension_worst",
+        ]
 
-    print('Downloading data from %s...' % url)
-    wis_df = pd.read_csv(url, header=None, names=df_cols, index_col=0)
-    wis_df.to_csv(data_file_path)
+        print('Downloading data from %s...' % url)
+        wis_df = pd.read_csv(url, header = None, names = df_cols, index_col = 0)
+        wis_df.to_csv(data_file_path)
+    else:
+        print(f"Re-using data file already downloaded to {data_file_path}.")
 
 
-def load_data(test_split=0.20):
+def load_data(test_split = 0.20):
     from sklearn.model_selection import train_test_split
     from sklearn.preprocessing import StandardScaler
 
@@ -79,7 +83,7 @@ def load_data(test_split=0.20):
     assert os.path.exists(data_file_path), \
         "%s - unable to open file!" % data_file_path
 
-    wis_df = pd.read_csv(data_file_path, index_col=0)
+    wis_df = pd.read_csv(data_file_path, index_col = 0)
     print(f"wis_df.shape: {wis_df.shape}")
 
     # diagnosis is the target col - char
@@ -88,133 +92,162 @@ def load_data(test_split=0.20):
     print(wis_df['diagnosis'].value_counts())
     # f_names = wis_df.columns[wis_df.columns != 'diagnosis']
 
-    X = wis_df.drop(['diagnosis'], axis=1).values
+    X = wis_df.drop(['diagnosis'], axis = 1).values
     y = wis_df['diagnosis'].values
     print(f"X.shape: {X.shape} - y.shape: {y.shape}")
 
     # split into train/test sets
     X_train, X_test, y_train, y_test = \
-        train_test_split(X, y, test_size=test_split,
-                         random_state=seed, stratify=y)
+        train_test_split(
+            X, y, test_size = test_split,
+            random_state = SEED, stratify = y
+        )
 
     # scale data
     ss = StandardScaler()
     X_train = ss.fit_transform(X_train)
     X_test = ss.transform(X_test)
-    # y_train = y_train[:, np.newaxis]
-    # y_test = y_test[:, np.newaxis]
+    y_train = y_train[:, np.newaxis]
+    y_test = y_test[:, np.newaxis]
 
     return (X_train, y_train), (X_test, y_test)
 
 
-# our ANN
+class WBCNet(nn.Module):
+    """ model for binary classification of Wisconsin Breast Cancer data """
 
-
-class WBCNet(pytk.PytkModule):
     def __init__(self, inp_size, num_classes):
         super(WBCNet, self).__init__()
-        # self.fc1 = nn.Linear(inp_size, 32)
-        # self.fc2 = nn.Linear(32, 32)
-        # self.out = nn.Linear(32, num_classes)
         self.net = nn.Sequential(
-            nn.Linear(inp_size, 16),
+            t3.Linear(inp_size, 16),
             nn.ReLU(),
-            nn.Dropout(p=0.10),
-            nn.Linear(16, 32),
+            nn.Dropout(p = 0.10),
+            t3.Linear(16, 32),
             nn.ReLU(),
-            nn.Dropout(p=0.10),
-            # nn.Linear(32, 64),
-            # nn.ReLU(),
-            # nn.Dropout(p=0.30),
-            nn.Linear(32, num_classes)
+            nn.Dropout(p = 0.10),
+            t3.Linear(32, num_classes),
+            nn.Sigmoid()
         )
 
     def forward(self, inp):
-        # x = F.relu(self.fc1(inp))
-        # x = F.relu(self.fc2(x))
-        # # x = F.sigmoid(self.out(x))
-        # x = self.out(x)
         x = self.net(inp)
         return x
 
 
 DO_TRAINING = True
-DO_TESTING = True
+DO_EVAL = True
 DO_PREDICTION = True
-MODEL_SAVE_PATH = './model_states/pyt_wbc_ann2.pyt'
+MODEL_SAVE_PATH = os.path.join(
+    os.path.dirname(__file__), 'model_states', 'pyt_wbc_ann2.pyt'
+)
 
 # Hyper-parameters
 NUM_FEATURES = 30
-NUM_CLASSES = 2
-NUM_EPOCHS = 750
+NUM_CLASSES = 1
+NUM_EPOCHS = 150
 BATCH_SIZE = 16
 LEARNING_RATE = 0.001
-METRICS_LIST = ['acc', 'f1', 'prec', 'rec']
-DECAY = 0.015
-
-
-def build_model():
-    model = WBCNet(NUM_FEATURES, NUM_CLASSES)
-    # define the loss function & optimizer that model should
-    # loss_fn = nn.BCELoss()
-    loss_fn = nn.CrossEntropyLoss()
-    optimizer = torch.optim.SGD(model.parameters(), lr=LEARNING_RATE, weight_decay=DECAY)
-    # optimizer = torch.optim.Adam(model.parameters(), lr=LEARNING_RATE, weight_decay=DECAY)
-    model.compile(loss=loss_fn, optimizer=optimizer, metrics=METRICS_LIST)
-    return model
+DECAY = 0.01
 
 
 def main():
+    # load our data & build tensor datasets for cross-training & testing
     (X_train, y_train), (X_test, y_test) = load_data()
-    # NOTE: BCELoss() functions expects labels to be floats (why can't it handle integers??)
-    # y_train, y_test = y_train.astype(np.float), y_test.astype(np.float)
     print(X_train.shape, y_train.shape, X_test.shape, y_test.shape)
-    # sys.exit(-1)
+    train_dataset = torch.utils.data.TensorDataset(
+        torch.FloatTensor(X_train),
+        torch.FloatTensor(y_train)
+    )
+    test_dataset = torch.utils.data.TensorDataset(
+        torch.FloatTensor(X_test),
+        torch.FloatTensor(y_test)
+    )
+
+    val_dataset, test_dataset = t3.split_dataset(
+        test_dataset, split_perc = 0.20
+    )
+    print(
+        f"train_dataset: {len(train_dataset)} - val_dataset: {len(val_dataset)} - " +
+        f"test_dataset: {len(test_dataset)}"
+    )
+
+    loss_fn = nn.BCELoss()
+    metrics_map = {
+        "acc": BinaryAccuracy(),
+        "f1": BinaryF1Score(),
+        "roc_auc": BinaryAUROC(thresholds = None)
+    }
 
     if DO_TRAINING:
-        print('Building model...')
-        model = build_model()
+        model = WBCNet(NUM_FEATURES, NUM_CLASSES)
+        optimizer = torch.optim.SGD(
+            model.parameters(), lr = LEARNING_RATE,
+            weight_decay = DECAY
+        )
         print(model)
 
         # train model
         print('Training model...')
-        # split training data into train/cross-val datasets in 80:20 ratio
-        hist = model.fit(X_train, y_train, validation_split=0.20,
-                         epochs=NUM_EPOCHS, batch_size=BATCH_SIZE, report_interval=50)
-        pytk.show_plots(hist, metric='f1', plot_title="Performance Metrics (f1-score)")
-
+        hist = t3.cross_train_model(
+            model, train_dataset, loss_fn, optimizer, device = DEVICE,
+            validation_dataset = val_dataset, epochs = NUM_EPOCHS,
+            batch_size = BATCH_SIZE, metrics_map = metrics_map,
+            reporting_interval = 25
+        )
+        hist.plot_metrics(fig_size = (16, 8))
         # save model state
-        model.save(MODEL_SAVE_PATH)
+        t3.save_model(model, MODEL_SAVE_PATH)
         del model
 
-    if DO_TESTING:
+    if DO_EVAL:
         # evaluate model performance on train/eval & test datasets
-        print('\nEvaluating model performance...')
-        model = build_model()
-        model.load(MODEL_SAVE_PATH)
+        model = WBCNet(NUM_FEATURES, NUM_CLASSES)
+        model = t3.load_model(model, MODEL_SAVE_PATH)
         print(model)
 
-        loss, acc, f1, prec, rec = model.evaluate(X_train, y_train, metrics=METRICS_LIST)
-        print(f'  Training dataset  -> loss: {loss:.4f} - acc: {acc:.4f} ' +
-              f'- f1: {f1:.4f} - prec: {prec:.4f} - rec: {rec:.4f}')
-        loss, acc, f1, prec, rec = model.evaluate(X_test, y_test, metrics=METRICS_LIST)
-        print(f'  Test dataset      -> loss: {loss:.4f} - acc: {acc:.4f} ' +
-              f'- f1: {f1:.4f} - prec: {prec:.4f} - rec: {rec:.4f}')
+        print('Evaluating model performance...')
+        print('Training dataset')
+        metrics = t3.evaluate_model(
+            model, train_dataset, loss_fn, device = DEVICE,
+            metrics_map = metrics_map,
+            batch_size = BATCH_SIZE
+        )
+        print(f"Training metrics: {metrics}")
+
+        print('Cross-validation dataset')
+        metrics = t3.evaluate_model(
+            model, val_dataset, loss_fn, device = DEVICE,
+            metrics_map = metrics_map,
+            batch_size = BATCH_SIZE
+        )
+        print(f"Cross-validation metrics: {metrics}")
+
+        print('Testing dataset')
+        metrics = t3.evaluate_model(
+            model, test_dataset, loss_fn, device = DEVICE,
+            metrics_map = metrics_map,
+            batch_size = BATCH_SIZE
+        )
+        print(f"Testing metrics: {metrics}")
         del model
 
     if DO_PREDICTION:
         print('\nRunning predictions...')
-        model = build_model()
-        model.load(MODEL_SAVE_PATH)
+        model = WBCNet(NUM_FEATURES, NUM_CLASSES)
+        model = t3.load_model(model, MODEL_SAVE_PATH)
         print(model)
 
-        # y_pred = np.round(model.predict(X_test)).reshape(-1)
-        y_pred = np.argmax(model.predict(X_test), axis=1)
-        # display output
-        print('Sample labels: ', y_test.flatten())
-        print('Sample predictions: ', y_pred)
-        print('We got %d/%d correct!' %
-              ((y_test.flatten() == y_pred).sum(), len(y_test.flatten())))
+        preds, actuals = t3.predict_dataset(model, test_dataset, device = DEVICE)
+        preds = np.round(preds).ravel()
+        actuals = actuals.ravel()
+        incorrect_counts = (preds != actuals).sum()
+        print(f"We got {incorrect_counts} of {len(actuals)} predictions wrong!")
+        print(classification_report(actuals, preds))
+        t3.plot_confusion_matrix(
+            confusion_matrix(actuals, preds),
+            class_names = ["Benign", "Malignant"],
+            title = "Cancer Prediction - Wisconsin Breast Cancer Data"
+        )
         del model
